almost all done

+++ SRC_HEAD/pipe_redirection/split_redir.c
@@ -6,7 +6,7 @@
 /*   By: oelazzou <oelazzou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/26 01:33:51 by oelazzou          #+#    #+#             */
-/*   Updated: 2020/12/30 18:34:23 by oelazzou         ###   ########.fr       */
+/*   Updated: 2020/12/30 18:38:02 by oelazzou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ static size_t		countall_arr(char **agg, size_t delime_len)
 	return (-1);
 }
 
-char				*redirection_varname(char ***arr, char *str, int *i)
+char				*redirection_varname(char ***arr, char *str, size_t *i)
 {
 	size_t			c_len;
 	char			**agg;
